requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
mysqlclient==2.2.1
gunicorn==21.2.0
python-dotenv==1.0.0
